Cleanup: Move components to root
