--- app/components/index.ts
@@ -1,4 +1,0 @@
-export * from "./footer";
-export * from "./general";
-export * from "./header";
-export * from "./sponsors";
