+++ components/index.ts
@@ -1,0 +1,9 @@
+export * from "./about";
+export * from "./client/partners-grid";
+export * from "./contact";
+export * from "./email-form";
+export * from "./footer";
+export * from "./general";
+export * from "./header";
+export * from "./partners";
+export * from "./sponsors";
