+++ partner-data.ts
